+++ src/ios/BrazePlugin.h
@@ -35,6 +35,7 @@
 - (void)setHomeCity:(CDVInvokedUrlCommand *)command;
 - (void)setPhoneNumber:(CDVInvokedUrlCommand *)command;
 - (void)setLanguage:(CDVInvokedUrlCommand *)command;
+- (void)setLastKnownLocation:(CDVInvokedUrlCommand *)command;
 
 - (void)setPushNotificationSubscriptionType:(CDVInvokedUrlCommand *)command;
 - (void)setEmailNotificationSubscriptionType:(CDVInvokedUrlCommand *)command;